fix(mongo): derive BSON types from codec lookup, not hardcoded map (FL-11)
Replace CODEC_TO_BSON_TYPE with codec lookup resolution via
codecLookup.get(codecId).targetTypes[0]. Fix mongoDoubleCodec
targetTypes from "float" to "double" (correct BSON type name).

Add codecLookup as optional parameter to interpreter and
deriveJsonSchema. Update all test files with mock codec lookup.

+++ packages/3-mongo-target/2-mongo-adapter/src/core/codecs.ts
@@ -28,7 +28,7 @@ export const mongoStringCodec = mongoCodec({
 
 export const mongoDoubleCodec = mongoCodec({
   typeId: MONGO_DOUBLE_CODEC_ID,
-  targetTypes: ['float'],
+  targetTypes: ['double'],
   traits: ['equality', 'order', 'numeric'],
   decode: (wire: number) => wire,
   encode: (value: number) => value,
